Do not need a dedicated component for the Save button in Setting pages so remove it. Move the button back to SettingPageTemplate.
