+++ Source/Plugins/Core/com.equella.core/js/tsrc/components/SettingPageTemplate.tsx
@@ -3,7 +3,19 @@ import { ReactNode } from "react";
 import { commonString } from "../util/commonstrings";
 import MessageInfo from "./MessageInfo";
 import { NavigationGuard } from "./NavigationGuard";
-import SaveButton from "./SaveButton";
+import { Button, makeStyles } from "@material-ui/core";
+import { Save } from "@material-ui/icons";
+
+const useStyles = makeStyles({
+  floatingButton: {
+    position: "fixed",
+    top: 0,
+    right: 0,
+    marginTop: "80px",
+    marginRight: "16px",
+    width: "calc(25% - 112px)",
+  },
+});
 
 export interface SettingPageTemplateProps {
   /**
@@ -45,10 +57,24 @@ const SettingPageTemplate = ({
   preventNavigation,
   children,
 }: SettingPageTemplateProps) => {
+  const classes = useStyles();
   return (
     <>
       {children}
-      <SaveButton onSave={onSave} saveButtonDisabled={saveButtonDisabled} />
+
+      <Button
+        id={"_saveButton"}
+        color={"primary"}
+        className={classes.floatingButton}
+        variant={"contained"}
+        size={"large"}
+        onClick={onSave}
+        aria-label={commonString.action.save}
+        disabled={saveButtonDisabled}
+      >
+        <Save />
+        {commonString.action.save}
+      </Button>
 
       <MessageInfo
         title={commonString.result.success}
